backend/requirements.txt
Django
djangorestframework
psycopg2
django-cors-headers
gunicorn
django-storages[s3]
Pillow
drf-nested-routers
django-filter
djoser
drf-spectacular
django-debug-toolbar